refact #144: DTO 폴더 삭제, Response 및 info로 분리
