+++ Project_Timer/Data/lagacy/NetworkResult.swift
